% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\docType{data}
\name{A}
\alias{A}
\title{A small random matrix}
\format{An object of class \code{matrix} with 7 rows and 7 columns.}
\usage{
A
}
\description{
A small random matrix used only to demonstrate the package's algorithms in the \code{examples} sections of the package documentation.
}
\keyword{datasets}
