% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contractor.R
\name{contractor}
\alias{contractor}
\title{Contract a tridiagonal matrix}
\usage{
contractor(A)
}
\arguments{
\item{A}{A tridiagonal matrix of order n x n.}
}
\value{
A tridiagonal matrix of order (n-1) x (n-1) with the same permanent as \code{A}.
}
\description{
An internal function employed by the BG method for computing the permanent.
}
\examples{
data(A)
contractor(A)
}
