\name{expperm-package}
\alias{expperm-package}
\alias{expperm}
\docType{package}
\title{\packageTitle{expperm}}
\description{\packageDescription{expperm}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{expperm}
  \packageIndices{expperm}

  The package serves primarily to demonstrate the algorithms described in the accompanying paper, which is currently under review.
  
  We include versions, which are as similar as reasonably possible, of algorithms written in both R and C++. The R code is intended to facilitate testing, modification and re-use of the code while the C++ code is intended to implement the algorithms most efficiently for application to real problems.
}
\author{
  \packageAuthor{expperm}
  
  Maintainer: \packageMaintainer{expperm}
}
\references{
  Powell B., Smith P.A. (2019). "Computing expectations and marginal likelihoods for permutations." (In Submission).
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\keyword{math}
\concept{linkage error}
\concept{permanent}
\concept{permutation}
% \seealso{
%   Optional links to other man pages
% }
% \examples{
%   ## Optional simple examples of the most important functions
%   ## Use \dontrun{} around code to be shown but not executed
% }
