% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gray.R
\name{gray}
\alias{gray}
\title{Gray code enumeration}
\usage{
gray(ind)
}
\arguments{
\item{ind}{A binary string.}
}
\value{
The next binary string in the Gray code sequence.
}
\description{
Computes the next binary string in a Gray code sequence using Knuth's Algorithm G.
}
\examples{
n<-4
indi<-rep(0,n)
allinds<-c()
for(i in 1:(2^n)){
  allinds<-rbind(allinds,indi)
  indi<-gray(indi)
}
allinds
}
