% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istridiagonal.R
\name{is.tridiagonal}
\alias{is.tridiagonal}
\title{Checking a matrix is tridiagonal}
\usage{
is.tridiagonal(A)
}
\arguments{
\item{A}{A matrix.}
}
\value{
A logical variable. \code{TRUE} if the \code{A} is tridiagonal, \code{FALSE} otherwise.
}
\description{
A function for checking whether a matrix is tridiagonal. The check is used before attempting to apply the BG method for computing the permanent, since the method is only applicable to tridiagonal matrices.
}
\examples{
data(A)
is.tridiagonal(A)
data(triA)
is.tridiagonal(triA)
}
