% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triA.R
\docType{data}
\name{triA}
\alias{triA}
\title{A small random tridiagonal matrix}
\format{An object of class \code{matrix} with 7 rows and 7 columns.}
\usage{
triA
}
\description{
A small random tridiagonal matrix used only to demonstrate the package's algorithms in the \code{examples} sections of the package documentation.
}
\keyword{datasets}
